Feat/tea 94 data validation test

+++ backend/src/controllers/AuthController.js
@@ -1,19 +1,27 @@
-import create from "../services/user/create.js";
-import UserRepo from "../database/repositories/UserRepo.js";
-import login from "../services/user/login.js";
-import ResetTokenRepo from "../database/repositories/ResetTokenRepo.js";
-import { NotFoundError } from "../lib/errors/index.js";
-import { sendResetLink } from "../services/user/password_reset.js";
-import EmailVerificationTokenRepo from "../database/repositories/emailVerificationRepo.js";
-import bcrypt from "bcrypt";
-import Joi from 'joi'
-import PasswordComplexity from "joi-password-complexity"
+import create from '../services/user/create.js';
+import UserRepo from '../database/repositories/UserRepo.js';
+import login from '../services/user/login.js';
+import ResetTokenRepo from '../database/repositories/ResetTokenRepo.js';
+import { NotFoundError } from '../lib/errors/index.js';
+import { sendResetLink } from '../services/user/password_reset.js';
+import EmailVerificationTokenRepo from '../database/repositories/emailVerificationRepo.js';
+import bcrypt from 'bcrypt';
+import Joi from 'joi';
+import PasswordComplexity from 'joi-password-complexity';
 
-
+import { validatePayload } from '../utils/index.js';
 
 export default class AuthController {
   static signup = async (req, res, next) => {
     try {
+      /**
+       * Validate Request
+       */
+      const errors = validatePayload(req);
+
+      // Update this latter
+      if (errors && Object.keys(errors).length > 0) throw errors;
+
       await create(req.body);
 
       res.status(201).json({
@@ -26,32 +34,39 @@ export default class AuthController {
     }
   };
 
-    static loginUser = async (req, res, next) => {
-        const body = req.body;
-        try {
-            const loggedInUser = await login(body, req, res);
-   
-            //set request cookie
-            if (loggedInUser.user && loggedInUser.token){
-                req.session.user = loggedInUser.user;
-                req.session.authorized = true;
+  static loginUser = async (req, res, next) => {
+    const body = req.body;
+    try {
+      /**
+       * Validate Request
+       */
+      const errors = validatePayload(req);
 
-                if (body.rememberMe) {
-                    req.session.user.maxAge = 30 * 24 * 60 * 60 * 1000; // Cookie expires after 30 days
-                } else {
-                    req.session.user.expires = false; // Cookie expires at end of session
-                }
-  
-                return res.send({ 
-                    message: "Logged in Successfully",
-                    user: loggedInUser.user,
-                    token: loggedInUser.token,
-                });  
-            }                   
-        } catch (error) {
-            next(error);
+      // Update this latter
+      if (errors && Object.keys(errors).length > 0) throw errors;
+      const loggedInUser = await login(body, req, res);
+
+      //set request cookie
+      if (loggedInUser.user && loggedInUser.token) {
+        req.session.user = loggedInUser.user;
+        req.session.authorized = true;
+
+        if (body.rememberMe) {
+          req.session.user.maxAge = 30 * 24 * 60 * 60 * 1000; // Cookie expires after 30 days
+        } else {
+          req.session.user.expires = false; // Cookie expires at end of session
         }
-    };
+
+        return res.send({
+          message: 'Logged in Successfully',
+          user: loggedInUser.user,
+          token: loggedInUser.token,
+        });
+      }
+    } catch (error) {
+      next(error);
+    }
+  };
 
   static logoutUser = async (req, res, next) => {
     try {
@@ -92,10 +107,18 @@ export default class AuthController {
     });
   };
 
-  static getResetLink = async (req, res) => {
+  static getResetLink = async (req, res, next) => {
     const { email } = req.body;
 
     try {
+      /**
+       * Validate Request
+       */
+      const errors = validatePayload(req);
+
+      // Update this latter
+      if (errors && Object.keys(errors).length > 0) throw errors;
+
       const registeredUser = await UserRepo.getUserByEmail(email);
 
       if (!registeredUser) throw new NotFoundError('Please input a valid registered email.');
@@ -108,62 +131,102 @@ export default class AuthController {
       return { message: 'A password reset link has been sent to your email address' };
     } catch (error) {
       console.error(error);
+      next(error);
     }
   };
 
   static updateUserPassword = async (req, res) => {
-    const { token, id, password } = req.body;
+    try {
+      const { token, id, password } = req.body;
 
-    await ResetTokenRepo.deleteExpiredTokens();
+      /**
+       * Validate Request
+       */
+      const errors = validatePayload(req);
 
-    const storedToken = await ResetTokenRepo.getToken(token, id);
+      // Update this latter
+      if (errors && Object.keys(errors).length > 0) throw errors;
 
-    if (!storedToken) {
-      throw new Error('Invalid or expired password reset token');
+      await ResetTokenRepo.deleteExpiredTokens();
+
+      const storedToken = await ResetTokenRepo.getToken(token, id);
+
+      if (!storedToken) {
+        throw new Error('Invalid or expired password reset token');
+      }
+
+      const isValid = await bcrypt.compare(token, storedToken.token);
+
+      if (!isValid) {
+        throw new Error('Invalid or expired password reset token');
+      }
+
+      const hashedPassword = await bcrypt.hash(password, Number(10));
+
+      const updatedUser = await UserRepo.updatePasswordById(id, hashedPassword);
+
+      if (!updatedUser) throw new NotFoundError('User not found');
+
+      res.status(200).redirect('/login');
+      return {
+        success: true,
+        message: 'Password has been updated successfully',
+        updatedUser,
+      };
+    } catch (error) {
+      next(error);
     }
+  };
 
-    const isValid = await bcrypt.compare(token, storedToken.token);
+  static verifyEmail = async (req, res, next) => {
+    try {
+      const errors = validatePayload(req);
 
-    if (!isValid) {
-      throw new Error('Invalid or expired password reset token');
+      // Update this latter
+      if (errors && Object.keys(errors).length > 0) throw errors;
+      const { token, id } = req.query;
+
+      await EmailVerificationTokenRepo.deleteExpiredTokens();
+
+      const storedToken = await EmailVerificationTokenRepo.getToken(id);
+
+      if (!storedToken) {
+        throw new Error('Invalid or expired email verification token');
+      }
+
+      const isValid = await bcrypt.compare(token, storedToken.token);
+
+      if (!isValid) {
+        throw new Error('Invalid or expired email verification token');
+      }
+
+      await UserRepo.updateById(id, { email_verified: 'true' });
+
+      return { message: 'email verified successfully' };
+    } catch (error) {
+      next(error);
     }
-
-    const hashedPassword = await bcrypt.hash(password, Number(10));
-
-    const updatedUser = await UserRepo.updatePasswordById(id, hashedPassword);
-
-    if (!updatedUser) throw new NotFoundError('User not found');
-
-    res.status(200).redirect('/login');
-    return {
-      success: true,
-      message: 'Password has been updated successfully',
-      updatedUser,
-    };
   };
 
   static updateUserPasswordFromMobile = async (req, res) => {
-
     try {
       // Validate with Joi
       const updateUserPassword = Joi.object({
-        id:Joi.string().required(),
+        id: Joi.string().required(),
         oldPassword: Joi.string().required(),
-        newPassword:new PasswordComplexity({
+        newPassword: new PasswordComplexity({
           min: 8,
           max: 25,
           lowerCase: 1,
           upperCase: 1,
           numeric: 1,
           symbol: 1,
-          requirementCount: 4
-        })
+          requirementCount: 4,
+        }),
       }).strict();
 
       if (updateUserPassword.validate(req.body).error) {
-        return res
-          .status(400)
-          .json(updateUserPassword.validate(req.body).error.details);
+        return res.status(400).json(updateUserPassword.validate(req.body).error.details);
       }
 
       // destruct request body
@@ -187,31 +250,7 @@ export default class AuthController {
       // Success
       return res.status(201).send('Password Successfully Updated');
     } catch (error) {
-      return res.status(500).send(`An error occurred whiles updating user password ${error}`)
+      return res.status(500).send(`An error occurred whiles updating user password ${error}`);
     }
-
-
-  }
-
-  static verifyEmail = async (req, res) => {
-    const { token, id } = req.query;
-
-    await EmailVerificationTokenRepo.deleteExpiredTokens();
-
-    const storedToken = await EmailVerificationTokenRepo.getToken(id);
-
-    if (!storedToken) {
-      throw new Error('Invalid or expired email verification token');
-    }
-
-    const isValid = await bcrypt.compare(token, storedToken.token);
-
-    if (!isValid) {
-      throw new Error('Invalid or expired email verification token');
-    }
-
-    await UserRepo.updateById(id, { email_verified: 'true' });
-
-    return { message: 'email verified successfully' };
   };
 }
